+++ python/dalex/dalex/datasets/__init__.py
@@ -1,5 +1,6 @@
-from ._load import load_titanic
+from ._load import load_titanic, load_fifa
 
 __all__ = [
-    "load_titanic"
+    "load_titanic",
+    "load_fifa"
 ]